Refactor: move custom validations for reservation into a module to reduce code in the model
